# Conflicts:
#	src/main/resources/sql/V001__init.sql

+++ src/main/java/com/mate/trackq/dao/ProjectDao.java
@@ -8,7 +8,7 @@ public interface ProjectDao {
 
     void create(Project project);
 
-    Project getById(Integer projectId);
+    Project getById(Long projectId);
 
     Project getByName(String projectName);
 }
